refactor: 💡 import path shortened
