+++ packages/tree/src/ATrees.ts
@@ -2,7 +2,7 @@ import { Objet } from '@jamashita/anden-object';
 import { BinaryPredicate, ForEach, Kind, Nullable } from '@jamashita/anden-type';
 import { Collection } from '@jamashita/lluvia-collection';
 import { ATree } from './ATree';
-import { ATreeNode } from './TreeNode/ATreeNode';
+import { ATreeNode } from './TreeNode';
 import { Trees } from './Trees';
 
 export abstract class ATrees<K, V, T extends ATreeNode<V, T>, E extends ATree<V, T>, C extends Collection<K, E>> extends Objet implements Trees<K, V, E> {
